requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.1
click==8.1.7
itsdangerous==2.1.2

PyPDF2==3.0.1
python-docx==0.8.11

nltk==3.8.1
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.24.3

requests==2.31.0
gunicorn==21.2.0
